avoid rotation to depend on head logic
* move state event into its own module
