+++ hydra-node/src/Hydra/API/Server.hs
@@ -32,9 +32,10 @@ import Hydra.Cardano.Api (LedgerEra)
 import Hydra.Chain (Chain (..))
 import Hydra.Chain.ChainState (IsChainState)
 import Hydra.Chain.Direct.State ()
-import Hydra.Events (EventSink (..), EventSource (..), StateEvent (..))
+import Hydra.Events (EventSink (..), EventSource (..))
 import Hydra.HeadLogic.Outcome qualified as StateChanged
 import Hydra.HeadLogic.State (Deposit (..), SeenSnapshot (..), seenSnapshotNumber)
+import Hydra.HeadLogic.StateEvent (StateEvent (..))
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Network (IP, PortNumber)
 import Hydra.Node.Environment (Environment)
