+++ hydra-node/src/Hydra/Events.hs
@@ -16,9 +16,6 @@ module Hydra.Events where
 import Hydra.Prelude
 
 import Conduit (ConduitT, MonadUnliftIO, ResourceT, runResourceT, sourceToList)
-import Hydra.Chain.ChainState (IsChainState)
-import Hydra.HeadLogic.Outcome (StateChanged)
-import Hydra.Tx.IsTx (ArbitraryIsTx)
 
 type EventId = Word64
 
@@ -37,8 +34,6 @@ newtype EventSource e m = EventSource
 getEvents :: (HasEventId e, MonadUnliftIO m) => EventSource e m -> m [e]
 getEvents EventSource{sourceEvents} = runResourceT $ sourceToList sourceEvents
 
-type LogId = EventId
-
 newtype EventSink e m = EventSink
   { putEvent :: HasEventId e => e -> m ()
   -- ^ Send a single event to the event sink.
@@ -50,29 +45,3 @@ putEventsToSinks sinks events =
   forM_ events $ \event ->
     forM_ sinks $ \sink ->
       putEvent sink event
-
--- * State change events as used by Hydra.Node
-
--- | A state change event with an event id that is the common entity to be
--- loaded from an 'EventSource' and sent to 'EventSink's.
-data StateEvent tx = StateEvent
-  { eventId :: EventId
-  , stateChanged :: StateChanged tx
-  , time :: UTCTime
-  }
-  deriving (Generic)
-
-instance HasEventId (StateEvent tx) where
-  getEventId = eventId
-
-deriving instance IsChainState tx => Show (StateEvent tx)
-deriving instance IsChainState tx => Eq (StateEvent tx)
-deriving instance IsChainState tx => ToJSON (StateEvent tx)
-deriving instance IsChainState tx => FromJSON (StateEvent tx)
-
-instance (ArbitraryIsTx tx, IsChainState tx) => Arbitrary (StateEvent tx) where
-  arbitrary = arbitrary >>= genStateEvent
-  shrink = genericShrink
-
-genStateEvent :: StateChanged tx -> Gen (StateEvent tx)
-genStateEvent sc = StateEvent <$> arbitrary <*> pure sc <*> arbitrary
