+++ hydra-node/test/Hydra/Events/FileBasedSpec.hs
@@ -9,10 +9,11 @@ import Hydra.Chain.Direct.State ()
 
 import Conduit (runConduitRes, sinkList, (.|))
 import Data.List (zipWith3)
-import Hydra.Events (EventSink (..), EventSource (..), StateEvent (..), getEvents, putEvent)
+import Hydra.Events (EventSink (..), EventSource (..), getEvents, putEvent)
 import Hydra.Events.FileBased (mkFileBasedEventStore)
 import Hydra.Events.Rotation (EventStore (..))
 import Hydra.HeadLogic (StateChanged)
+import Hydra.HeadLogic.StateEvent (StateEvent (..))
 import Hydra.Ledger.Cardano (Tx)
 import Hydra.Ledger.Simple (SimpleTx)
 import Hydra.Persistence (PersistenceIncremental (..), createPersistenceIncremental)
